Finishing the guided setup script
